# InternPool: begin conversion to thread-safe data structure