--- a/lib/std/Thread.zig
+++ b/lib/std/Thread.zig
@@ -280,12 +280,13 @@ pub fn getCurrentId() Id {
 pub const CpuCountError = error{
     PermissionDenied,
     SystemResources,
+    Unsupported,
     Unexpected,
 };
 
 /// Returns the platforms view on the number of logical CPU cores available.
 pub fn getCpuCount() CpuCountError!usize {
-    return Impl.getCpuCount();
+    return try Impl.getCpuCount();
 }
 
 /// Configuration options for hints on how to spawn threads.
@@ -780,6 +781,10 @@ const WasiThreadImpl = struct {
 
     fn getCurrentId() Id {
         return tls_thread_id;
+    }
+
+    fn getCpuCount() error{Unsupported}!noreturn {
+        return error.Unsupported;
     }
 
     fn getHandle(self: Impl) ThreadHandle {