This is another step towards first putting codegen on a separate thread from sema and later making both sema and codegen multi-threaded.  Currently, the multi-threaded features of the new data structure are disabled by `InternPool.want_multi_threaded` because they have a large performance impact on the compiler and are not yet in use.  It is still undecided how to minimize the impact of 

The main conflicting change is replacing `*Zcu` with `Zcu.PerThread` in any code that wants to be able to mutate the intern pool.  I chose to pass this information around everywhere instead of attempting something with `threadlocal` variables because in the fast case that would turn the intern pool into a global singleton, and in the slow case would introduce an extra lookup on every intern pool mutation.  Interestingly, if there is a desire to remove intern pool mutation from the backends, the per thread change can be reverted over time in parts of the code and that guarantees that the mutations can no longer in that section.

The current performance impact is in the range where I expect to be able to recoup the loss with more work, because while timing individual commits, this is close to the same time as before the change which ended up having the worst performance impact, which I mitigated by disabling features of the data structure as mentioned above.  Even with this current slowdown, there may be a desire to get some or all of this change merged sooner to reduce future conflicts, and continue work separately.
```
Benchmark 1 (28 runs): master/bin/zig build-exe -ODebug --dep aro --dep aro_translate_c --dep build_options -Mroot=src/main.zig -Maro=lib/compiler/aro/aro.zig --dep aro -Maro_translate_c=lib/compiler/aro_translate_c.zig -Mbuild_options=options.zig -fno-emit-bin
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          4.43s  ± 69.4ms    4.28s  … 4.52s           1 ( 4%)        0%
  peak_rss            323MB ± 1.06MB     320MB …  325MB          1 ( 4%)        0%
  cpu_cycles         23.1G  ±  233M     22.8G  … 23.5G           0 ( 0%)        0%
  instructions       45.0G  ± 34.3K     45.0G  … 45.0G           2 ( 7%)        0%
  cache_references   2.13G  ± 18.6M     2.11G  … 2.19G           2 ( 7%)        0%
  cache_misses       86.9M  ± 1.15M     84.5M  … 89.7M           2 ( 7%)        0%
  branch_misses      85.7M  ±  577K     84.6M  … 87.1M           2 ( 7%)        0%
Benchmark 2 (26 runs): tsip/bin/zig build-exe -ODebug --dep aro --dep aro_translate_c --dep build_options -Mroot=src/main.zig -Maro=lib/compiler/aro/aro.zig --dep aro -Maro_translate_c=lib/compiler/aro_translate_c.zig -Mbuild_options=options.zig -fno-emit-bin
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          4.68s  ± 54.9ms    4.50s  … 4.74s           5 (19%)        💩+  5.7% ±  0.8%
  peak_rss            324MB ±  682KB     323MB …  325MB          0 ( 0%)          +  0.4% ±  0.2%
  cpu_cycles         23.3G  ±  168M     23.1G  … 23.6G           0 ( 0%)          +  0.9% ±  0.5%
  instructions       40.1G  ± 32.4K     40.1G  … 40.1G           1 ( 4%)        ⚡- 11.0% ±  0.0%
  cache_references   2.15G  ± 15.4M     2.13G  … 2.20G           2 ( 8%)         +   1.2% ±  0.4%
  cache_misses       79.2M  ± 1.55M     77.0M  … 82.8M           0 ( 0%)        ⚡-  8.9% ±  0.9%
  branch_misses      73.4M  ±  564K     72.6M  … 74.8M           0 ( 0%)        ⚡- 14.3% ±  0.4%
```