@andrewrk 2024-07-08
Data point: building the zig compiler with the x86_64 backend with codegen threading enabled:

```diff
--- a/src/InternPool.zig
+++ b/src/InternPool.zig
@@ -93,7 +93,7 @@ files: std.AutoArrayHashMapUnmanaged(Cache.BinDigest, OptionalDeclIndex) = .{},
 /// Whether a multi-threaded intern pool is useful.
 /// Currently `false` until the intern pool is actually accessed
 /// from multiple threads to reduce the cost of this data structure.
-const want_multi_threaded = false;
+const want_multi_threaded = true;
 
 /// Whether a single-threaded intern pool impl is in use.
 pub const single_threaded = builtin.single_threaded or !want_multi_threaded;
diff --git a/src/target.zig b/src/target.zig
index 2accc100b8..e236cea616 100644
--- a/src/target.zig
+++ b/src/target.zig
@@ -572,6 +572,7 @@ pub inline fn backendSupportsFeature(backend: std.builtin.CompilerBackend, compt
             else => false,
         },
         .separate_thread => switch (backend) {
+            .stage2_x86_64 => true,
             else => false,
         },
     };
```

before: 8f20e81b8816aadd8ceb1b04bd3727cc1d124464
after: ab4eeb770a0af411d525616dcdbfa4a8491f8ac0
(before/after this PR merged)

```
Benchmark 1 (3 runs): /home/andy/dev/zig/build-release/stage3/bin/zig build-exe -fno-llvm -fno-lld --stack 33554432 -fno-sanitize-thread -ODebug --dep aro --dep aro_translate_c --dep build_options -Mroot=/home/andy/dev/zig/src/main.zig -Maro=/home/andy/dev/zig/lib/compiler/aro/aro.zig --dep aro -Maro_translate_c=/home/andy/dev/zig/lib/compiler/aro_translate_c.zig -Mbuild_options=/home/andy/dev/zig/.zig-cache/c/ecf36f9c39c1cbfb3043b07b8688debd/options.zig --cache-dir /home/andy/dev/zig/.zig-cache --global-cache-dir /home/andy/.cache/zig --name zig
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          12.8s  ±  189ms    12.7s  … 13.1s           0 ( 0%)        0%
  peak_rss            403MB ±  463KB     403MB …  404MB          0 ( 0%)        0%
  cpu_cycles         64.3G  ±  124M     64.2G  … 64.5G           0 ( 0%)        0%
  instructions        133G  ±  171K      133G  …  133G           0 ( 0%)        0%
  cache_references   4.07G  ± 4.88M     4.06G  … 4.07G           0 ( 0%)        0%
  cache_misses        419M  ± 3.33M      416M  …  423M           0 ( 0%)        0%
  branch_misses       435M  ±  831K      434M  …  436M           0 ( 0%)        0%
Benchmark 2 (3 runs): /home/andy/src/zig/build-release/stage3/bin/zig build-exe -fno-llvm -fno-lld --stack 33554432 -fno-sanitize-thread -ODebug --dep aro --dep aro_translate_c --dep build_options -Mroot=/home/andy/src/zig/src/main.zig -Maro=/home/andy/src/zig/lib/compiler/aro/aro.zig --dep aro -Maro_translate_c=/home/andy/src/zig/lib/compiler/aro_translate_c.zig -Mbuild_options=/home/andy/src/zig/.zig-cache/c/f18315ac6d459d8944cdba2c81e890df/options.zig --cache-dir /home/andy/src/zig/.zig-cache --global-cache-dir /home/andy/.cache/zig --name zig
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          8.57s  ± 66.9ms    8.52s  … 8.64s           0 ( 0%)        ⚡- 33.3% ±  2.5%
  peak_rss            517MB ± 3.73MB     513MB …  520MB          0 ( 0%)        💩+ 28.1% ±  1.5%
  cpu_cycles         65.3G  ±  131M     65.2G  … 65.4G           0 ( 0%)        💩+  1.5% ±  0.4%
  instructions        138G  ± 1.22M      138G  …  138G           0 ( 0%)        💩+  3.5% ±  0.0%
  cache_references   4.01G  ± 17.6M     4.00G  … 4.03G           0 ( 0%)          -  1.4% ±  0.7%
  cache_misses        220M  ± 4.17M      215M  …  223M           0 ( 0%)        ⚡- 47.5% ±  2.0%
  branch_misses       340M  ± 2.84M      338M  …  343M           0 ( 0%)        ⚡- 21.9% ±  1.1%

```