@Jarred-Sumner 2024-07-09
wow 12.8s end-to-end before

with `-fno-emit-bin`, bun currently takes ~11s (it's about 1m 15s end-to-end, most of which is the "Emit LLVM" step)

```ts
❯ time .cache/zig/zig build check --summary new --verbose
info: zig compiler v0.13.0
/root/bun/.cache/zig/zig build-obj -freference-trace=16 -fno-strip -fno-stack-check -fno-stack-protector -fno-omit-frame-pointer -fvalgrind -fPIC -ODebug -target native-native-gnu.2.27 -mcpu native --dep async_io --dep zlib-internal --dep async --dep ZigGeneratedClasses --dep ResolvedSourceTag --dep build_options -Mroot=/root/bun/root.zig -Masync_io=/root/bun/src/io/io_linux.zig -Mzlib-internal=/root/bun/src/deps/zlib.posix.zig -Masync=/root/bun/src/async/posix_event_loop.zig -MZigGeneratedClasses=/root/bun/build/codegen/ZigGeneratedClasses.zig -MResolvedSourceTag=/root/bun/build/codegen/ResolvedSourceTag.zig -Mbuild_options=/root/bun/.zig-cache/c/e16e8fad5a41dcf3d3a150b03d8ff201/options.zig -lc++ -lc -fno-emit-bin -fformatted-panics --eh-frame-hdr --emit-relocs -ffunction-sections --cache-dir /root/bun/.zig-cache --global-cache-dir /root/.cache/zig --name bun-debug -fno-compiler-rt --zig-lib-dir /root/bun/src/deps/zig/lib --listen=-
Build Summary: 3/3 steps succeeded
check success
└─ zig build-obj bun-debug Debug native-native-gnu.2.27 success 11s MaxRSS:563M
   └─ options success

________________________________________________________
Executed in   14.32 secs    fish           external
   usr time   14.19 secs    0.00 micros   14.19 secs
   sys time    0.74 secs  482.00 micros    0.74 secs
```