@andrewrk 2024-07-09
Note that my 12.8s data point above is with the x86_64 backend. You can try enabling it, but it's not the default yet due to machine code quality, debug info correctness, and behavior test coverage.

https://github.com/ziglang/zig/blob/854e86c5676de82bc46b5c13a0c9c807596e438d/build.zig#L218-L220

If you're on an M1 you'll have to wait for the aarch64 backend to get a similar speedup.